+++ mythtv/libs/libmythui/mythmainwindow.cpp
@@ -2864,12 +2864,12 @@ QSize MythMainWindow::NormSize(const QSize &size)
 
 int MythMainWindow::NormX(const int x)
 {
-    return (int)(qRound(x * d->wmult));
+    return qRound(x * d->wmult);
 }
 
 int MythMainWindow::NormY(const int y)
 {
-    return (int)(qRound(y * d->hmult));
+    return qRound(y * d->hmult);
 }
 
 void MythMainWindow::SetScalingFactors(float wmult, float hmult)
